+++ src/theme/theme.types.ts
@@ -322,6 +322,28 @@ export type ThemeColors = {
   disabledBackground: string;
   /** gray3 */
   disabledText: string;
+
+  /** light: gray800, dark: white */
+  dropDownButtonText: string;
+  /** light: blue500, dark: blue400 */
+  dropDownButtonTextHover: string;
+  /** light: blue500, dark: blue400 */
+  dropDownButtonTextActive: string;
+  /** light: gray400, dark: gray600 */
+  dropDownButtonTextDisabled: string;
+  /** light: gray100, dark: gray700 */
+  dropDownButtonBackground: string;
+  /** light: gray100, dark: gray700 */
+  dropDownButtonBackgroundHover: string;
+  /** light: blue100, dark: blue800 */
+  dropDownButtonBackgroundActive: string;
+  /** light: gray200, dark: gray900 */
+  dropDownButtonBackgroundDisabled: string;
+  /** light: gray200, dark: gray800 */
+  dropDownButtonBackgroundLoading: string;
+  /** light: red500, dark: red500 */
+  dropDownButtonBorderError: string;
+
   /** gray6 */
   divider: string;
   /** gray4 */
@@ -416,6 +438,10 @@ export type ThemeColors = {
   negative: string;
   /** negative or brandPink */
   negativeBlackBackground: string;
+  /** green600, pink600, blue500, teal600 */
+  newsLabelTextColors: string[];
+  /** green100, pink100, blue100, teal100 */
+  newsLabelBackgroundColors: string[];
   /** brandBlue */
   onboardingAccentBlue: string;
   /** brandGreen */